savegame/legacy: move to trx
