--- src/tr2/game/savegame.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/savegame.h>
