+++ src/tr2/meson.build
@@ -89,7 +89,6 @@ endif
 
 sources = [
   init,
-  'game/savegame/savegame_legacy.c',
   resources,
 ]
